Fix for TestSegmentsChecksumError()
Change to throw IOException instead of Exception. The code in FindSegmentsFile.Run expects the reads to throw an IOException when the file is corrupt or in some other way invalid like failing the crc32 check. Based on this exception type, the code will ignore the last segment and continue. This is what the unit test is looking for.