Better but not a full fix for TestThreadInterruptDeadlock and TestTwoThreadsInterruptDeadlock.
C# already has a ThreadingInterruptedException and core components like SyncTextWriter throw this exception when they are interrupted. I updated the unit test to catch both Lucene.Net.Util.ThreadInterruptedException (already did this) and System.Threading.ThreadInterruptedException.

This change makes both of the unit tests fail during the first phase of the two phase commit because they are interrupted. Looking into the IndexWriter class, this seems to be the expected behavior but the unit test still says it is not good. This needs more research.