Moving to a separate branch for independent PR.
The ParallelCompositComposit reader now has a ParallelCompositReaderAnonymousInnerClassHelper for the inner instance instead of ParallelCompositReader. Looking back at the history, it has been this way for a long time but this test does not exist in the java version. Is this test in place to show there is something that needs to be fixed in ParallelCompositComposit reader?