Fix for TestThreadInterruptDeadlock
Use the C# ThreadInterruptedException so the catches will properly control the flow during interruption.
This fixes the two unit tests Lucene.Next.Index.TestIndexWriter.TestThreadInterruptDeadlock and Lucene.Next.Index.TestIndexWriter.TestTwoThreadsInterruptDeadlock
Lucene uses the ThreadInterruptedException for code flow control so it is important to use the C# ThreadInterruptedException to avoid duplicating the catch everywhere and preserve the flow.