Use concrete ConcurrentHashSet instead of abstraction for performance (#960)
* Use concrete ConcurrentHashSet instead of abstraction for performance

Calling into an interface when its not necessary will result in poorer performance because the JIT needs to do some type analysis. When it is a concrete type, this is avoided. Since this is an internal field there's no reason to use the abstracted type. 

See https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/quality-rules/ca1859